+++ docs/CHANGELOG.rst
@@ -12,6 +12,11 @@ Fixed
 -----
 - Raise an error if a user calls the download method
   on a data object and inputs a non-existing file or field name
+Changed
+-----
+- Change the parsing logic for the one-to-one metadata table
+  in ``BaseTables`` class so it does not fail when the table 
+  index does not reflect the state of samples in the collection
 
 
 ===================